users/serializers.py
from rest_framework import serializers

from users.models import CustomUser


class CustomUserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = CustomUser
        fields = ["url", "name",]
        #  "age",]
